support wallet rpc listening and binding to internal wallet rpc process
update to monero-project v0.17.1.9
support maximum sync period for wallets
cmakelists finds system openssl and boost
update readme architecture diagram, documentation section
update api spec
change nonce from Integer to Long
support daemon waitForNextBlockHeader()
support daemon info isBusySyncing(), isSynchronized()
support zmq in rpc connection and wallet rpc listener (unused)
txs can become confirmed or unlocked during merging
fix bug merging txs when output newly confirmed
update junit4 to junit5
add TestAll.java which tests classes in same order as javascript
rename to WalletTxTracker and support waitForUnlockedBalance() for tests

+++ src/main/java/monero/daemon/model/MoneroBlock.java
@@ -237,7 +237,7 @@ public class MoneroBlock extends MoneroBlockHeader {
   }
   
   @Override
-  public MoneroBlock setNonce(Integer nonce) {
+  public MoneroBlock setNonce(Long nonce) {
     super.setNonce(nonce);
     return this;
   }
